chore(language-js): separate unit, integration, and performance tests
