+++ packages/language-js/src/__tests__/integration/index.test.ts
@@ -6,7 +6,7 @@ import type {
 
 import { expect } from "chai";
 
-import JavaScriptLanguagePlugin from "../index";
+import JavaScriptLanguagePlugin from "../../index";
 
 suite("The @webmangler/language-js plugin", function() {
   test("no argument", function() {
