+++ packages/language-js/src/expressions/__tests__/common/benchmark-helpers.ts
