+++ packages/language-js/src/expressions/__tests__/unit/common.test.ts
@@ -3,7 +3,7 @@ import { expect } from "chai";
 import {
   sampleValues,
   selectorCombinators,
-} from "../values";
+} from "../common";
 
 import { patterns } from "../../common";
